Refactored and rearanged code
