+++ resources/ts/Modifiers/Modifier.ts
