--- resources/ts/classes/Adders/SearchIconAdder.ts
@@ -1,19 +1,0 @@
-import conf from '@/conf'
-import AdderInterface from '@/classes/Adders/Adder'
-
-export default class implements AdderInterface {
-    public add(): void {
-        const target = document.querySelector(conf.selectors.targetForSearchIcon)
-
-        if (!target)
-            return
-
-        const html = `
-            <svg xmlns="http://www.w3.org/2000/svg" class="php-revival-search-icon" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" />
-            </svg>
-        `
-
-        target.insertAdjacentHTML('afterbegin', html)
-    }
-}
