+++ resources/ts/templates/cardItemTemplate.ts
@@ -1,7 +1,7 @@
-import type { CardItem } from '@/types'
+import type { RandomVideo } from '@/types'
 import conf from '@/conf'
 
-export default (card: CardItem): string => {
+export default (card: RandomVideo): string => {
     return `
         <a href="${card.link}" class="revival-random-video" target="_blank">
             <img src="${conf.urls.server}${card.img}" alt="${card.title}">
